% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_dissolve.R
\name{st_truchet_dissolve}
\alias{st_truchet_dissolve}
\title{Dissolving the boundaries of individual tiles in Truchet mosaics}
\usage{
st_truchet_dissolve(mosaic)
}
\arguments{
\item{mosaic}{a mosaic produced by function \code{st_truchet_ms}}
}
\value{
An object of type \code{sf} with the mosaic after dissolving the boundaries of individual tiles
}
\description{
Dissolving the boundaries of individual tiles in Truchet mosaics
}
\examples{
mosaic <- st_truchet_ms()
mosaic <- st_truchet_dissolve(mosaic)
}
