% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_flex.R
\name{st_truchet_flex}
\alias{st_truchet_flex}
\title{Flexible Truchet tiles}
\usage{
st_truchet_flex(x = 0, y = 0, type = "Al", b = 1/2)
}
\arguments{
\item{x}{A number with the x coordinate of the center of the tile}

\item{y}{A number with the y coordinate of the center of the tile}

\item{type}{A single character to designate a type of tile; currently supported options are "Ac", "Bc", "Cc", "Dc", "As", "Bs", "Cs", "Ds"}

\item{b}{A number between zero and one that controls the shape of the boundary between the two parts of the tile}
}
\value{
A list with one or more objects of type \code{sf} representing one or more tiles depending on type
}
\description{
Flexible Truchet tiles
}
\note{
For a discussion of Truchet patterns see: Robert Bosch & Urchin Colley (2013) Figurative mosaics from flexible Truchet tiles, Journal of Mathematics and the Arts, 7:3-4, 122-135, \url{10.1080/17513472.2013.838830}
}
\examples{
st_truchet_flex(type = "Al")
st_truchet_flex(type = "Cl")
}
