% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_fm.R
\name{st_truchet_fm}
\alias{st_truchet_fm}
\title{Mosaics with flexible Truchet tiles}
\usage{
st_truchet_fm(
  df = NULL,
  tiles = c("Al", "Cl"),
  b = 1/2,
  xlim = c(1, 3),
  ylim = c(1, 6)
)
}
\arguments{
\item{df}{an (optional) data frame with the following columns: x and y (the coordinates of the tiles in a 1 by 1 grid), tiles (characters with types of tiles to use for mosaic), b (control of the boundary; defaults to 1/2)}

\item{tiles}{a character vector with types of tiles to use for mosaic (default: \code{c("dr", "dl")})}

\item{b}{A number between zero and one that controls the shape of the boundary between the two parts of the tile}

\item{xlim}{a numeric vector of length 2 giving the range of the x coordinates of the mosaic (ignored if argument \code{df} is an input)}

\item{ylim}{a numeric vector of length 2 giving the range of the y coordinates of the mosaic (ignored if argument \code{df} is an input)}
}
\value{
An object of type \code{sf} with the tiles arranged as a mosaic
}
\description{
Mosaics with flexible Truchet tiles
}
\note{
For a discussion of Truchet patterns see \url{http://arearugscarpet.blogspot.com/2014/04/the-curse-of-truchets-tiles.html}
}
\examples{
mosaic <- st_truchet_fm()
plot(mosaic)
mosaic <- st_truchet_fm(b = 1/3)
plot(mosaic)
}
