% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_l.R
\name{st_truchet_l}
\alias{st_truchet_l}
\title{Truchet tiles made with polygons}
\usage{
st_truchet_l(x = 0, y = 0, type = "dl")
}
\arguments{
\item{x}{A number with the x coordinate of the center of the tile}

\item{y}{A number with the y coordinate of the center of the tile}

\item{type}{A single character to designate a type of tile; currently supported options are "dl", "dr"}
}
\value{
A list with one or more objects of type \code{sf} representing one or more tiles depending on type
}
\description{
Truchet tiles made with polygons
}
\note{
For a discussion of Truchet patterns see \url{http://arearugscarpet.blogspot.com/2014/04/the-curse-of-truchets-tiles.html}
}
\examples{
st_truchet_l(type = "dl")
st_truchet_l(type = "dr")
}
