% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_ms.R
\name{st_truchet_ms}
\alias{st_truchet_ms}
\title{Truchet mosaics}
\usage{
st_truchet_ms(t1, t2 = NULL, prop = 0.5, xlim = c(1, 8), ylim = c(1, 12))
}
\arguments{
\item{t1}{a \code{sf} data frame with truchet tile(s) produced by \link{st_truchet_p} at scale 1; these tiles are squares with sides of length 1}

\item{t2}{an (optional) \code{sf} data frame with truchet tile(s) produced by \link{st_truchet_p} at scale 1/2; these tiles are squares with sides of length 1/2}

\item{prop}{a number between 0 and 1 with the proportion of spots in the mosaic to cover with tiles of scale 1}

\item{xlim}{a numeric vector of length 2 giving the range of the x coordinates of the mosaic}

\item{ylim}{a numeric vector of length 2 giving the range of the y coordinates of the mosaic}
}
\value{
A list with two objects of type \code{sf} representing the container for the mosaic and the tiles arranged as a mosaic
}
\description{
Truchet mosaics
}
\note{
For a discussion of multi-scale Truchet patterns see \url{https://christophercarlson.com/portfolio/multi-scale-truchet-patterns/}
}
\examples{
tiles_1 <- st_truchet_p(type = "-", scale_p = 1)
tiles_2 <- st_truchet_p(type = "f", scale_p = 1/2)
mosaic <- st_truchet_ms(tiles_1, tiles_2)
}
