% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_ss.R
\name{st_truchet_ss}
\alias{st_truchet_ss}
\title{Truchet mosaics}
\usage{
st_truchet_ss(df = NULL, tiles = c("dr", "dl"), xlim = c(1, 3), ylim = c(1, 6))
}
\arguments{
\item{df}{an (optional) data frame with the following columns: x and y (the coordinates of the tiles in a 1 by 1 grid), tiles (characters with types of tiles to use for mosaic), scale_p (the scale of the tile to be placed at each coordinate)}

\item{tiles}{a character vector with types of tiles to use for mosaic (default: \code{c("dr", "dl")})}

\item{xlim}{a numeric vector of length 2 giving the range of the x coordinates of the mosaic (ignored if argument \code{df} is an input)}

\item{ylim}{a numeric vector of length 2 giving the range of the y coordinates of the mosaic (ignored if argument \code{df} is an input)}
}
\value{
An object of type \code{sf} with the tiles arranged as a mosaic
}
\description{
Truchet mosaics
}
\note{
For a discussion of multi-scale Truchet patterns see \url{https://christophercarlson.com/portfolio/multi-scale-truchet-patterns/}
}
\examples{
mosaic <- st_truchet_ss()
mosaic <- st_truchet_ss(tiles = c("dl", "dr"))
}
