% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truchet-package.R
\docType{package}
\name{truchet-package}
\alias{truchet}
\alias{truchet-package}
\title{truchet: Tiles and Random Mosaics Truchet Style Using the 'sf' Framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to create Truchet tiles, so called after Sébastien Truchet who was the first to describe the patterns obtained by rotating tiles with respect to each other. This form of tiling is described by Smith and Boucher (1987) \url{https://muse.jhu.edu/article/600574}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paezha.github.io/truchet/}
  \item \url{https://github.com/paezha/truchet}
  \item Report bugs at \url{https://github.com/paezha/truchet/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}

}
\keyword{internal}
