% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_boutique.R
\name{st_truchet_boutique}
\alias{st_truchet_boutique}
\title{Flexible Truchet tiles}
\usage{
st_truchet_boutique(x = 0, y = 0, type = "ribbon_1")
}
\arguments{
\item{x}{A number with the x coordinate of the center of the tile}

\item{y}{A number with the y coordinate of the center of the tile}

\item{type}{A single character to designate a type of tile; currently supported options are "ribbon_1", "ribbon_2", "ribbon_3", "ribbon_4", "paradise_1", "paradise_2", "paradise_3", "paradise_4", "silk_1", "silk_2", "silk_3", "silk_4", "rainbow_1", "rainbow_2", "cloud_1", "cloud_2", "cloud_3", "cloud_4"}
}
\value{
A data frame with one or more objects of type \code{sf} representing one or more tiles depending on type
}
\description{
Flexible Truchet tiles
}
\note{
For a discussion of variable tiling patterns see: Robert J.Krawczyk (2020) Infinitely Variable Tiling Patterns: From Truchet to Sol LeWitt Revisited, Patterns, 1:5, 1-4, \url{10.1016/j.patter.2020.100084} and
Robert J.Krawczyk (2011) Truchet tilings revisited, Proceedings of ISAMA 2011, 69-77 \url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.378.5320&rep=rep1&type=pdf#page=69}
}
\examples{
st_truchet_boutique(type = "ribbon_1")
st_truchet_boutique(type = "ribbon_2")
}
