% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_truchet_p.R
\name{st_truchet_p}
\alias{st_truchet_p}
\title{Truchet tiles made with polygons}
\usage{
st_truchet_p(type = "d", scale_p = 1)
}
\arguments{
\item{type}{A single character to designate a type of tile; currently supported options are "d", "-", "f", "a"}

\item{scale_p}{A number to designate the scale of the tile; currently supported options are 1 and 1/2}
}
\value{
A list with one or more objects of type \code{sf} representing one or more tiles depending on type
}
\description{
Truchet tiles made with polygons
}
\note{
For a discussion of multi-scale Truchet patterns see \url{https://christophercarlson.com/portfolio/multi-scale-truchet-patterns/}
}
\examples{
st_truchet_p(type = "-")
st_truchet_p(type = "f", scale_p = 1/2)
}
