% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truchet-package.R
\docType{package}
\name{truchet-package}
\alias{truchet}
\alias{truchet-package}
\title{truchet: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paezha.github.io/truchet/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
