% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truchet-package.R
\docType{package}
\name{truchet-package}
\alias{truchet}
\alias{truchet-package}
\title{truchet: Package for creating Truchet tiles and random mosaics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package includes utility functions to create Truchet tiles and arrange them to create mosaics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paezha.github.io/truchet/}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
